Implementacion solicitud/generar extracto de cuenta. correciones de variables y formatos. actualizacion filtros de tablas por fechas. correcion de longitudes de variables

+++ pkgInterfaz/formActualizarCliente.cs
@@ -14,10 +14,10 @@ namespace testForms.pkgInterfaz
     public partial class formActualizarCliente : Form
     {
 
-        int id_usuarioActual = 0;
+        long id_usuarioActual = 0;
         string clave;
         Datos data = new Datos();
-        public formActualizarCliente(int prm_usuarioActual)
+        public formActualizarCliente(long prm_usuarioActual)
         {
             InitializeComponent();
             id_usuarioActual = prm_usuarioActual;
